requirements.txt
ultralytics>=8.0.0
deep-sort-realtime>=1.3.2
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
twilio>=8.5.0
pandas>=2.0.0
requests>=2.31.0